% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateIMPACT.R
\name{aggregateIMPACT}
\alias{aggregateIMPACT}
\title{Function to aggregate IMPACT results at a specified level}
\usage{
aggregateIMPACT(df = NULL, level = "regglo", aggr_type = "sum")
}
\arguments{
\item{df}{Specific IMPACT results as a list object. Likely output of readGDX}

\item{level}{"reg" Short for regional. "regglo" for regional+global
result being passed on to this function}

\item{aggr_type}{sum or mean}
}
\value{
Country or FPU level aggregation of IMPACT results
}
\description{
Function to aggregate IMPACT results at a specified level
}
\examples{
\dontrun{
aggregateIMPACT()
}
}
\author{
Abhijeet Mishra
}
