% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMapping.R
\name{findMapping}
\alias{findMapping}
\title{Function to get IMPACT mapping for crops, activities, regions(from country)
This function is just a glorified file reader as you can provide any mapping
you want. See description below.}
\usage{
findMapping(name = NULL, file = "default_mappings.xlsx", multimap = FALSE, ...)
}
\arguments{
\item{name}{Name of mapping to get.
You can provide any name here but the options when using default mapping are
\itemize{
 \item{"c"}{ --> Crops}
 \item{"j"}{ --> Activities}
 \item{"cty"}{ --> Countries with 6 CGIAR region mappings}
}
If using the "multimap" switch for this function. Your mapping MUST have the
first column named "type" which should contain the same text as "name"
parameter of this function. For example, if the "type" column is filled with
"blabla" then the "name" parameter here should also be called "blabla".
In default example, for crops, the "type" column is filled with "c" in the
default file.}

\item{file}{Mapping file in excel format. Uses the file provided with this
package by default.}

\item{multimap}{If the mapping file contains multiple mappings in single
file. See the "default_mappings.xlsx" file provided with this package for
example. There, single excel file contains "j", "c" and "cty" mappings. You
have to stick to naming conventions described above if you activate this
switch. Defaults to FALSE. Set it to TRUE if using multiple mapping in 1 file
If "multimap" is set to TRUE, your mapping MUST have the
first column named "type" which should contain the same text as "name"
parameter of this function. For example, if the "type" column is filled with
"blabla" then the "name" parameter here should also be called "blabla".
In default example, for crops, the "type" column is filled with "c" in the
default file.}

\item{...}{All other parameters to be passed on to read_xlsx function for
reading the excel file of mapping using read_xlsx. See ?read_xlsx.}
}
\description{
Function to get IMPACT mapping for crops, activities, regions(from country)
This function is just a glorified file reader as you can provide any mapping
you want. See description below.
}
\examples{
\dontrun{
getMapping(name = "c")
}
}
\author{
Abhijeet Mishra
}
