% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDX.R
\name{readGDX}
\alias{readGDX}
\title{R utility package for reading objects from a GDX file.}
\usage{
readGDX(
  gdx,
  name,
  use_model_name = "IMPACT",
  verbosity = FALSE,
  quick_df = FALSE
)
}
\arguments{
\item{gdx}{Path to the gdx file}

\item{name}{name of the parameter being read}

\item{use_model_name}{Which model name should be shown. Use "IMPACT" for
IMPACT results.}

\item{verbosity}{If additional messages should be printed about variable
being read}

\item{quick_df}{if description and model name column should be dropped}
}
\value{
Parameter and domain of gdx output result queried
}
\description{
R utility package for reading objects from a GDX file.
}
\examples{
\dontrun{
readGDX()
}
}
\author{
Abhijeet Mishra
}
