% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_get_domain_mapping.R
\name{tool_get_domain_mapping}
\alias{tool_get_domain_mapping}
\title{Function which pulls the domain list}
\usage{
tool_get_domain_mapping(domain = NULL, file = "mapping_items.xlsx")
}
\arguments{
\item{domain}{set}

\item{file}{Mapping file}
}
\value{
Mapping object for IMPACT framework
}
\description{
Function which pulls the domain list
}
\examples{
\dontrun{
tool_get_mapping()
}
}
\author{
Abhijeet Mishra
}
