% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_get_mapping.R
\name{tool_get_mapping}
\alias{tool_get_mapping}
\title{tool_get_mapping}
\usage{
tool_get_mapping(type = "cty")
}
\arguments{
\item{type}{Defaults to "cty" for country mapping. Choose "fpu" for
IMPACT FPU mapping.}
}
\value{
Mapping object for IMPACT framework
}
\description{
tool_get_mapping
}
\examples{
\dontrun{
tool_get_mapping()
}
}
\author{
Abhijeet Mishra
}
