% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateIMPACT.R
\name{aggregateIMPACT}
\alias{aggregateIMPACT}
\title{aggregateIMPACT}
\usage{
aggregateIMPACT(df = NULL, level = "reg", aggr_type = "sum", weight = NULL)
}
\arguments{
\item{df}{list object from IMPACT results. Likely output of readGDX}

\item{level}{"reg" Short for regional.
result being passed on to this function}

\item{aggr_type}{sum}

\item{weight}{Weight by which aggregation can be made}
}
\value{
Country or FPU level aggregation of IMPACT results
}
\description{
aggregateIMPACT
}
\examples{
\dontrun{
aggregateIMPACT()
}
}
\author{
Abhijeet Mishra
}
