% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat_source_build.R
\name{drat_source_build}
\alias{drat_source_build}
\title{Build package utility for DRAT linkage of IFPRI R packages}
\usage{
drat_source_build(lib = ".")
}
\arguments{
\item{lib}{Path to library}
}
\description{
Build package utility for DRAT linkage of IFPRI R packages
}
\examples{
\dontrun{
drat_source_build()
}
}
\author{
Abhijeet Mishra
}
