% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDX.R
\name{readGDX}
\alias{readGDX}
\title{readGDX}
\usage{
readGDX(gdx, name, use_model_name = "IMPACT")
}
\arguments{
\item{gdx}{Path to the gdx file}

\item{name}{name of the parameter being read}

\item{use_model_name}{Which model name should be shown. Use "IMPACT" for
IMPACT results.}
}
\value{
Parameter and domain of gdx output result queried
}
\description{
readGDX
}
\examples{
\dontrun{
readGDX()
}
}
\author{
Abhijeet Mishra
}
