% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_get_mapping.R
\name{tool_get_mapping}
\alias{tool_get_mapping}
\title{Function for getting correct mapping of IFPRI regions (CTY to regions)}
\usage{
tool_get_mapping(
  type = NULL,
  file = "mapping_items.xlsx",
  sheet = NULL,
  sp_mapping = "Standard-IMPACT_dis1"
)
}
\arguments{
\item{type}{Defaults to "cty" for country mapping. Choose "fpu" for
IMPACT FPU mapping.}

\item{file}{Mapping file}

\item{sheet}{Sheet from which data is read. Ideally output of
tool_get_domain_mapping function}

\item{sp_mapping}{Column from Aggregation Regions sheet}
}
\value{
Mapping object for IMPACT framework
}
\description{
Function for getting correct mapping of IFPRI regions (CTY to regions)
}
\examples{
\dontrun{
tool_get_mapping()
}
}
\author{
Abhijeet Mishra
}
