% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_namespace_description.R
\name{update_namespace_description}
\alias{update_namespace_description}
\title{Simple update for latest dependencies}
\usage{
update_namespace_description(pkg = ".")
}
\arguments{
\item{pkg}{Package directory}
}
\description{
Simple update for latest dependencies
}
\examples{
\dontrun{
update_namespace_description()
}
}
\author{
Abhijeet Mishra
}
